Refactor the style chooser package.
- Merged and removed some classes (IO and IOexample).
- Renamed some classes for better meaning.
- Removed unused classes.
